refactor: move login/logout button file
